9.5.19:33

+++ src/fashion_server/SomeFunction.java
@@ -60,24 +60,24 @@ public class SomeFunction {
     			
     			String pure, grid, dot, floral, cross_stripe, vertical_stripe, number_letter, repeat, favor_pattern;
     			favor_pattern = "pattern_pure";
-    			int pure_int, grid_int, dot_int, floral_int, cross_stripe_int, vertical_stripe_int, number_letter_int, repeat_int, favor_int;
+    			float pure_int, grid_int, dot_int, floral_int, cross_stripe_int, vertical_stripe_int, number_letter_int, repeat_int, favor_int;
     			if(rs.next()) {
     				pure = rs.getString("pure");
-    				pure_int = Integer.parseInt(pure);
+    				pure_int = Float.parseFloat(pure);
     				grid = rs.getString("grid");
-    				grid_int = Integer.parseInt(grid);
+    				grid_int = Float.parseFloat(grid);
     				dot = rs.getString("dot");
-    				dot_int = Integer.parseInt(dot);
+    				dot_int = Float.parseFloat(dot);
     				floral = rs.getString("floral");
-    				floral_int = Integer.parseInt(floral);
+    				floral_int = Float.parseFloat(floral);
     				cross_stripe = rs.getString("cross_stripe");
-    				cross_stripe_int = Integer.parseInt(cross_stripe);
+    				cross_stripe_int = Float.parseFloat(cross_stripe);
     				vertical_stripe = rs.getString("vertical_stripe");
-    				vertical_stripe_int = Integer.parseInt(vertical_stripe);
+    				vertical_stripe_int = Float.parseFloat(vertical_stripe);
     				number_letter = rs.getString("number_letter");
-    				number_letter_int = Integer.parseInt(number_letter);
+    				number_letter_int = Float.parseFloat(number_letter);
     				repeat = rs.getString("repeat");
-    				repeat_int = Integer.parseInt(repeat);
+    				repeat_int = Float.parseFloat(repeat);
     				
     				favor_int = pure_int;
     				if(grid_int > favor_int) {
@@ -117,14 +117,14 @@ public class SomeFunction {
     			
     			String sleeve_sleeveless, sleeve_short, sleeve_long, favor_sleeve;
     			favor_sleeve = "sleeve_length_sleeveless";
-    			int sleeve_sleeveless_int, sleeve_short_int, sleeve_long_int, favor_sleeve_int;
+    			float sleeve_sleeveless_int, sleeve_short_int, sleeve_long_int, favor_sleeve_int;
     			if(rs.next()) {
     				sleeve_sleeveless = rs.getString("sleeveless");
-    				sleeve_sleeveless_int = Integer.parseInt(sleeve_sleeveless);
+    				sleeve_sleeveless_int = Float.parseFloat(sleeve_sleeveless);
     				sleeve_short = rs.getString("short");
-    				sleeve_short_int = Integer.parseInt(sleeve_short);
+    				sleeve_short_int = Float.parseFloat(sleeve_short);
     				sleeve_long = rs.getString("long");
-    				sleeve_long_int = Integer.parseInt(sleeve_long);
+    				sleeve_long_int = Float.parseFloat(sleeve_long);
     				
     				favor_sleeve_int = sleeve_sleeveless_int;
     				if(sleeve_short_int > favor_sleeve_int) {
@@ -144,14 +144,14 @@ public class SomeFunction {
     			
     			String tight, straight, loose, favor_model;
     			favor_model = "model_tight";
-    			int tight_int, straight_int, loose_int, favor_model_int;
+    			float tight_int, straight_int, loose_int, favor_model_int;
     			if(rs.next()) {
     				tight = rs.getString("tight");
-    				tight_int = Integer.parseInt(tight);
+    				tight_int = Float.parseFloat(tight);
     				straight = rs.getString("straight");
-    				straight_int = Integer.parseInt(straight);
+    				straight_int = Float.parseFloat(straight);
     				loose = rs.getString("loose");
-    				loose_int = Integer.parseInt(loose);
+    				loose_int = Float.parseFloat(loose);
     				
     				favor_model_int = tight_int;
     				if(straight_int > favor_model_int) {
@@ -171,14 +171,14 @@ public class SomeFunction {
     			
     			String cloth_short, cloth_mid, cloth_long, favor_length;
     			favor_length = "length_short";
-    			int cloth_short_int, cloth_mid_int, cloth_long_int, favor_length_int;
+    			float cloth_short_int, cloth_mid_int, cloth_long_int, favor_length_int;
     			if(rs.next()) {
     				cloth_short = rs.getString("short");
-    				cloth_short_int = Integer.parseInt(cloth_short);
+    				cloth_short_int = Float.parseFloat(cloth_short);
     				cloth_mid = rs.getString("mid");
-    				cloth_mid_int = Integer.parseInt(cloth_mid);
+    				cloth_mid_int = Float.parseFloat(cloth_mid);
     				cloth_long = rs.getString("long");
-    				cloth_long_int = Integer.parseInt(cloth_long);
+    				cloth_long_int = Float.parseFloat(cloth_long);
     				
     				favor_length_int = cloth_short_int;
     				if(cloth_mid_int > favor_length_int) {
@@ -198,22 +198,22 @@ public class SomeFunction {
     			
     			String stand, Vshape, bateau, round, lapel, high, hoodie, favor_collar_shape;
     			favor_collar_shape = "collar_shape_stand";
-    			int stand_int, Vshape_int, bateau_int, round_int, lapel_int, high_int, hoodie_int, favor_collar_shape_int;
+    			float stand_int, Vshape_int, bateau_int, round_int, lapel_int, high_int, hoodie_int, favor_collar_shape_int;
     			if(rs.next()) {
     				stand = rs.getString("stand");
-    				stand_int = Integer.parseInt(stand);
+    				stand_int = Float.parseFloat(stand);
     				Vshape = rs.getString("Vshape");
-    				Vshape_int = Integer.parseInt(Vshape);
+    				Vshape_int = Float.parseFloat(Vshape);
     				bateau = rs.getString("bateau");
-    				bateau_int = Integer.parseInt(bateau);
+    				bateau_int = Float.parseFloat(bateau);
     				round = rs.getString("round");
-    				round_int = Integer.parseInt(round);
+    				round_int = Float.parseFloat(round);
     				lapel = rs.getString("lapel");
-    				lapel_int = Integer.parseInt(lapel);
+    				lapel_int = Float.parseFloat(lapel);
     				high = rs.getString("high");
-    				high_int = Integer.parseInt(high);
+    				high_int = Float.parseFloat(high);
     				hoodie = rs.getString("hoodie");
-    				hoodie_int = Integer.parseInt(hoodie);
+    				hoodie_int = Float.parseFloat(hoodie);
     				
     				favor_collar_shape_int = stand_int;
     				if(Vshape_int > favor_collar_shape_int) {
